+++ spring-native-aws-lambda-function/src/main/java/com/coffeebeans/springnativeawslambda/Application.java
@@ -16,19 +16,17 @@
  *
  */
 
-package com.coffeebeans.springnativeawslambda.function;
+package com.coffeebeans.springnativeawslambda;
 
-import com.coffeebeans.springnativeawslambda.function.model.Request;
-import com.coffeebeans.springnativeawslambda.function.model.Response;
-import org.springframework.aot.hint.RuntimeHints;
-import org.springframework.aot.hint.RuntimeHintsRegistrar;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ImportRuntimeHints;
 
-public class AppRuntimeHints implements RuntimeHintsRegistrar {
+@SpringBootApplication
+@ImportRuntimeHints(AppRuntimeHints.class)
+public class Application {
 
-  @Override
-  public void registerHints(RuntimeHints hints, ClassLoader classLoader) {
-    hints.serialization().registerType(Request.class);
-    hints.serialization().registerType(Response.class);
-    hints.resources().registerPattern("org/joda/time/tz/*");
+  public static void main(String[] args) {
+    SpringApplication.run(Application.class, args);
   }
 }
