+++ spring-native-aws-lambda-function/src/main/java/com/coffeebeans/springnativeawslambda/model/Request.java
@@ -16,12 +16,9 @@
  *
  */
 
-package com.coffeebeans.springnativeawslambda.function.model;
+package com.coffeebeans.springnativeawslambda.model;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonProperty.Access;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import java.io.Serializable;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -32,13 +29,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Response implements Serializable {
+public class Request implements Serializable {
 
   @NotBlank
-  @JsonProperty(access = Access.READ_ONLY)
   private String name;
-
-  @NotNull
-  @JsonProperty(access = Access.READ_ONLY)
-  private boolean saved;
 }
